https://github.com/Monkestation/MonkeStation/pull/609 needs auxgm supermatter first

+++ code/modules/atmospherics/environmental/LINDA_fire.dm
@@ -161,7 +161,7 @@
 	if((temperature < FIRE_MINIMUM_TEMPERATURE_TO_EXIST) || (volume <= 1))
 		qdel(src)
 		return
-	if(!location.air || location.air.get_oxidation_power() < 0.5 || (INSUFFICIENT(GAS_PLASMA) && INSUFFICIENT(GAS_TRITIUM) && location.air.get_fuel_amount() < 0.5))
+	if(!location.air || location.air.get_oxidation_power() < 0.5 || (INSUFFICIENT(GAS_PLASMA) && INSUFFICIENT(GAS_TRITIUM && INSUFFICIENT(GAS_H2)) && location.air.get_fuel_amount() < 0.5))
 		qdel(src)
 		return
